Refactor status and model modules
